refactor out utils and clean up debug logging add actual logging
